int helyett float

+++ teglalap.py
@@ -1,3 +1,3 @@
-a = int(input("a: "))
-b = int(input("b: "))
+a = float(input("a: "))
+b = float(input("b: "))
 print("Terulet:", a*b)
